BAH-4494 | Kanchi | fix. Review comments.
