+++ packages/bahmni-services/src/userService/index.ts
@@ -3,9 +3,8 @@ export {
   getUserLoginLocation,
   getAvailableLocations,
   getDefaultDateFormat,
-  logout,
   saveUserLocation,
   updateSessionLocation,
 } from './userService';
 export { type User, type UserLocation } from './models';
-export { BAHMNI_USER_LOCATION_COOKIE } from './constants';
+export { BAHMNI_USER_LOCATION_COOKIE } from '../constants/app';
